feat: 덱추가부분 모달 추가, 뒤로가기 설정
