+++ src/assets/icons/CloseButton.jsx
@@ -1,0 +1,21 @@
+export const CloseButton = ({ onClick }) => {
+  return (
+    <svg
+      width='24'
+      height='24'
+      onClick={onClick}
+      style={{ cursor: 'pointer' }}
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      stroke='currentColor'
+    >
+      <path
+        strokeLinecap='round'
+        strokeLinejoin='round'
+        strokeWidth={2}
+        d='M6 18L18 6M6 6l12 12'
+      />
+    </svg>
+  );
+};
